Merge data frames

Merge data frames

rbind_df(list_df)

Arguments

list_df
A list of multiple data frames with the same variables in the same order.

Value

One merged data frame with the names of the first input data frame.

Examples

rbind_df(list(iris, iris))

#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1            5.1         3.5          1.4         0.2     setosa
#> 2            4.9         3.0          1.4         0.2     setosa
#> 3            4.7         3.2          1.3         0.2     setosa
#> 4            4.6         3.1          1.5         0.2     setosa
#> 5            5.0         3.6          1.4         0.2     setosa
#> 6            5.4         3.9          1.7         0.4     setosa
#> 7            4.6         3.4          1.4         0.3     setosa
#> 8            5.0         3.4          1.5         0.2     setosa
#> 9            4.4         2.9          1.4         0.2     setosa
#> 10           4.9         3.1          1.5         0.1     setosa
#> 11           5.4         3.7          1.5         0.2     setosa
#> 12           4.8         3.4          1.6         0.2     setosa
#> 13           4.8         3.0          1.4         0.1     setosa
#> 14           4.3         3.0          1.1         0.1     setosa
#> 15           5.8         4.0          1.2         0.2     setosa
#> 16           5.7         4.4          1.5         0.4     setosa
#> 17           5.4         3.9          1.3         0.4     setosa
#> 18           5.1         3.5          1.4         0.3     setosa
#> 19           5.7         3.8          1.7         0.3     setosa
#> 20           5.1         3.8          1.5         0.3     setosa
#> 21           5.4         3.4          1.7         0.2     setosa
#> 22           5.1         3.7          1.5         0.4     setosa
#> 23           4.6         3.6          1.0         0.2     setosa
#> 24           5.1         3.3          1.7         0.5     setosa
#> 25           4.8         3.4          1.9         0.2     setosa
#> 26           5.0         3.0          1.6         0.2     setosa
#> 27           5.0         3.4          1.6         0.4     setosa
#> 28           5.2         3.5          1.5         0.2     setosa
#> 29           5.2         3.4          1.4         0.2     setosa
#> 30           4.7         3.2          1.6         0.2     setosa
#> 31           4.8         3.1          1.6         0.2     setosa
#> 32           5.4         3.4          1.5         0.4     setosa
#> 33           5.2         4.1          1.5         0.1     setosa
#> 34           5.5         4.2          1.4         0.2     setosa
#> 35           4.9         3.1          1.5         0.2     setosa
#> 36           5.0         3.2          1.2         0.2     setosa
#> 37           5.5         3.5          1.3         0.2     setosa
#> 38           4.9         3.6          1.4         0.1     setosa
#> 39           4.4         3.0          1.3         0.2     setosa
#> 40           5.1         3.4          1.5         0.2     setosa
#> 41           5.0         3.5          1.3         0.3     setosa
#> 42           4.5         2.3          1.3         0.3     setosa
#> 43           4.4         3.2          1.3         0.2     setosa
#> 44           5.0         3.5          1.6         0.6     setosa
#> 45           5.1         3.8          1.9         0.4     setosa
#> 46           4.8         3.0          1.4         0.3     setosa
#> 47           5.1         3.8          1.6         0.2     setosa
#> 48           4.6         3.2          1.4         0.2     setosa
#> 49           5.3         3.7          1.5         0.2     setosa
#> 50           5.0         3.3          1.4         0.2     setosa
#> 51           7.0         3.2          4.7         1.4 versicolor
#> 52           6.4         3.2          4.5         1.5 versicolor
#> 53           6.9         3.1          4.9         1.5 versicolor
#> 54           5.5         2.3          4.0         1.3 versicolor
#> 55           6.5         2.8          4.6         1.5 versicolor
#> 56           5.7         2.8          4.5         1.3 versicolor
#> 57           6.3         3.3          4.7         1.6 versicolor
#> 58           4.9         2.4          3.3         1.0 versicolor
#> 59           6.6         2.9          4.6         1.3 versicolor
#> 60           5.2         2.7          3.9         1.4 versicolor
#> 61           5.0         2.0          3.5         1.0 versicolor
#> 62           5.9         3.0          4.2         1.5 versicolor
#> 63           6.0         2.2          4.0         1.0 versicolor
#> 64           6.1         2.9          4.7         1.4 versicolor
#> 65           5.6         2.9          3.6         1.3 versicolor
#> 66           6.7         3.1          4.4         1.4 versicolor
#> 67           5.6         3.0          4.5         1.5 versicolor
#> 68           5.8         2.7          4.1         1.0 versicolor
#> 69           6.2         2.2          4.5         1.5 versicolor
#> 70           5.6         2.5          3.9         1.1 versicolor
#> 71           5.9         3.2          4.8         1.8 versicolor
#> 72           6.1         2.8          4.0         1.3 versicolor
#> 73           6.3         2.5          4.9         1.5 versicolor
#> 74           6.1         2.8          4.7         1.2 versicolor
#> 75           6.4         2.9          4.3         1.3 versicolor
#> 76           6.6         3.0          4.4         1.4 versicolor
#> 77           6.8         2.8          4.8         1.4 versicolor
#> 78           6.7         3.0          5.0         1.7 versicolor
#> 79           6.0         2.9          4.5         1.5 versicolor
#> 80           5.7         2.6          3.5         1.0 versicolor
#> 81           5.5         2.4          3.8         1.1 versicolor
#> 82           5.5         2.4          3.7         1.0 versicolor
#> 83           5.8         2.7          3.9         1.2 versicolor
#> 84           6.0         2.7          5.1         1.6 versicolor
#> 85           5.4         3.0          4.5         1.5 versicolor
#> 86           6.0         3.4          4.5         1.6 versicolor
#> 87           6.7         3.1          4.7         1.5 versicolor
#> 88           6.3         2.3          4.4         1.3 versicolor
#> 89           5.6         3.0          4.1         1.3 versicolor
#> 90           5.5         2.5          4.0         1.3 versicolor
#> 91           5.5         2.6          4.4         1.2 versicolor
#> 92           6.1         3.0          4.6         1.4 versicolor
#> 93           5.8         2.6          4.0         1.2 versicolor
#> 94           5.0         2.3          3.3         1.0 versicolor
#> 95           5.6         2.7          4.2         1.3 versicolor
#> 96           5.7         3.0          4.2         1.2 versicolor
#> 97           5.7         2.9          4.2         1.3 versicolor
#> 98           6.2         2.9          4.3         1.3 versicolor
#> 99           5.1         2.5          3.0         1.1 versicolor
#> 100          5.7         2.8          4.1         1.3 versicolor
#> 101          6.3         3.3          6.0         2.5  virginica
#> 102          5.8         2.7          5.1         1.9  virginica
#> 103          7.1         3.0          5.9         2.1  virginica
#> 104          6.3         2.9          5.6         1.8  virginica
#> 105          6.5         3.0          5.8         2.2  virginica
#> 106          7.6         3.0          6.6         2.1  virginica
#> 107          4.9         2.5          4.5         1.7  virginica
#> 108          7.3         2.9          6.3         1.8  virginica
#> 109          6.7         2.5          5.8         1.8  virginica
#> 110          7.2         3.6          6.1         2.5  virginica
#> 111          6.5         3.2          5.1         2.0  virginica
#> 112          6.4         2.7          5.3         1.9  virginica
#> 113          6.8         3.0          5.5         2.1  virginica
#> 114          5.7         2.5          5.0         2.0  virginica
#> 115          5.8         2.8          5.1         2.4  virginica
#> 116          6.4         3.2          5.3         2.3  virginica
#> 117          6.5         3.0          5.5         1.8  virginica
#> 118          7.7         3.8          6.7         2.2  virginica
#> 119          7.7         2.6          6.9         2.3  virginica
#> 120          6.0         2.2          5.0         1.5  virginica
#> 121          6.9         3.2          5.7         2.3  virginica
#> 122          5.6         2.8          4.9         2.0  virginica
#> 123          7.7         2.8          6.7         2.0  virginica
#> 124          6.3         2.7          4.9         1.8  virginica
#> 125          6.7         3.3          5.7         2.1  virginica
#> 126          7.2         3.2          6.0         1.8  virginica
#> 127          6.2         2.8          4.8         1.8  virginica
#> 128          6.1         3.0          4.9         1.8  virginica
#> 129          6.4         2.8          5.6         2.1  virginica
#> 130          7.2         3.0          5.8         1.6  virginica
#> 131          7.4         2.8          6.1         1.9  virginica
#> 132          7.9         3.8          6.4         2.0  virginica
#> 133          6.4         2.8          5.6         2.2  virginica
#> 134          6.3         2.8          5.1         1.5  virginica
#> 135          6.1         2.6          5.6         1.4  virginica
#> 136          7.7         3.0          6.1         2.3  virginica
#> 137          6.3         3.4          5.6         2.4  virginica
#> 138          6.4         3.1          5.5         1.8  virginica
#> 139          6.0         3.0          4.8         1.8  virginica
#> 140          6.9         3.1          5.4         2.1  virginica
#> 141          6.7         3.1          5.6         2.4  virginica
#> 142          6.9         3.1          5.1         2.3  virginica
#> 143          5.8         2.7          5.1         1.9  virginica
#> 144          6.8         3.2          5.9         2.3  virginica
#> 145          6.7         3.3          5.7         2.5  virginica
#> 146          6.7         3.0          5.2         2.3  virginica
#> 147          6.3         2.5          5.0         1.9  virginica
#> 148          6.5         3.0          5.2         2.0  virginica
#> 149          6.2         3.4          5.4         2.3  virginica
#> 150          5.9         3.0          5.1         1.8  virginica
#> 151          5.1         3.5          1.4         0.2     setosa
#> 152          4.9         3.0          1.4         0.2     setosa
#> 153          4.7         3.2          1.3         0.2     setosa
#> 154          4.6         3.1          1.5         0.2     setosa
#> 155          5.0         3.6          1.4         0.2     setosa
#> 156          5.4         3.9          1.7         0.4     setosa
#> 157          4.6         3.4          1.4         0.3     setosa
#> 158          5.0         3.4          1.5         0.2     setosa
#> 159          4.4         2.9          1.4         0.2     setosa
#> 160          4.9         3.1          1.5         0.1     setosa
#> 161          5.4         3.7          1.5         0.2     setosa
#> 162          4.8         3.4          1.6         0.2     setosa
#> 163          4.8         3.0          1.4         0.1     setosa
#> 164          4.3         3.0          1.1         0.1     setosa
#> 165          5.8         4.0          1.2         0.2     setosa
#> 166          5.7         4.4          1.5         0.4     setosa
#> 167          5.4         3.9          1.3         0.4     setosa
#> 168          5.1         3.5          1.4         0.3     setosa
#> 169          5.7         3.8          1.7         0.3     setosa
#> 170          5.1         3.8          1.5         0.3     setosa
#> 171          5.4         3.4          1.7         0.2     setosa
#> 172          5.1         3.7          1.5         0.4     setosa
#> 173          4.6         3.6          1.0         0.2     setosa
#> 174          5.1         3.3          1.7         0.5     setosa
#> 175          4.8         3.4          1.9         0.2     setosa
#> 176          5.0         3.0          1.6         0.2     setosa
#> 177          5.0         3.4          1.6         0.4     setosa
#> 178          5.2         3.5          1.5         0.2     setosa
#> 179          5.2         3.4          1.4         0.2     setosa
#> 180          4.7         3.2          1.6         0.2     setosa
#> 181          4.8         3.1          1.6         0.2     setosa
#> 182          5.4         3.4          1.5         0.4     setosa
#> 183          5.2         4.1          1.5         0.1     setosa
#> 184          5.5         4.2          1.4         0.2     setosa
#> 185          4.9         3.1          1.5         0.2     setosa
#> 186          5.0         3.2          1.2         0.2     setosa
#> 187          5.5         3.5          1.3         0.2     setosa
#> 188          4.9         3.6          1.4         0.1     setosa
#> 189          4.4         3.0          1.3         0.2     setosa
#> 190          5.1         3.4          1.5         0.2     setosa
#> 191          5.0         3.5          1.3         0.3     setosa
#> 192          4.5         2.3          1.3         0.3     setosa
#> 193          4.4         3.2          1.3         0.2     setosa
#> 194          5.0         3.5          1.6         0.6     setosa
#> 195          5.1         3.8          1.9         0.4     setosa
#> 196          4.8         3.0          1.4         0.3     setosa
#> 197          5.1         3.8          1.6         0.2     setosa
#> 198          4.6         3.2          1.4         0.2     setosa
#> 199          5.3         3.7          1.5         0.2     setosa
#> 200          5.0         3.3          1.4         0.2     setosa
#> 201          7.0         3.2          4.7         1.4 versicolor
#> 202          6.4         3.2          4.5         1.5 versicolor
#> 203          6.9         3.1          4.9         1.5 versicolor
#> 204          5.5         2.3          4.0         1.3 versicolor
#> 205          6.5         2.8          4.6         1.5 versicolor
#> 206          5.7         2.8          4.5         1.3 versicolor
#> 207          6.3         3.3          4.7         1.6 versicolor
#> 208          4.9         2.4          3.3         1.0 versicolor
#> 209          6.6         2.9          4.6         1.3 versicolor
#> 210          5.2         2.7          3.9         1.4 versicolor
#> 211          5.0         2.0          3.5         1.0 versicolor
#> 212          5.9         3.0          4.2         1.5 versicolor
#> 213          6.0         2.2          4.0         1.0 versicolor
#> 214          6.1         2.9          4.7         1.4 versicolor
#> 215          5.6         2.9          3.6         1.3 versicolor
#> 216          6.7         3.1          4.4         1.4 versicolor
#> 217          5.6         3.0          4.5         1.5 versicolor
#> 218          5.8         2.7          4.1         1.0 versicolor
#> 219          6.2         2.2          4.5         1.5 versicolor
#> 220          5.6         2.5          3.9         1.1 versicolor
#> 221          5.9         3.2          4.8         1.8 versicolor
#> 222          6.1         2.8          4.0         1.3 versicolor
#> 223          6.3         2.5          4.9         1.5 versicolor
#> 224          6.1         2.8          4.7         1.2 versicolor
#> 225          6.4         2.9          4.3         1.3 versicolor
#> 226          6.6         3.0          4.4         1.4 versicolor
#> 227          6.8         2.8          4.8         1.4 versicolor
#> 228          6.7         3.0          5.0         1.7 versicolor
#> 229          6.0         2.9          4.5         1.5 versicolor
#> 230          5.7         2.6          3.5         1.0 versicolor
#> 231          5.5         2.4          3.8         1.1 versicolor
#> 232          5.5         2.4          3.7         1.0 versicolor
#> 233          5.8         2.7          3.9         1.2 versicolor
#> 234          6.0         2.7          5.1         1.6 versicolor
#> 235          5.4         3.0          4.5         1.5 versicolor
#> 236          6.0         3.4          4.5         1.6 versicolor
#> 237          6.7         3.1          4.7         1.5 versicolor
#> 238          6.3         2.3          4.4         1.3 versicolor
#> 239          5.6         3.0          4.1         1.3 versicolor
#> 240          5.5         2.5          4.0         1.3 versicolor
#> 241          5.5         2.6          4.4         1.2 versicolor
#> 242          6.1         3.0          4.6         1.4 versicolor
#> 243          5.8         2.6          4.0         1.2 versicolor
#> 244          5.0         2.3          3.3         1.0 versicolor
#> 245          5.6         2.7          4.2         1.3 versicolor
#> 246          5.7         3.0          4.2         1.2 versicolor
#> 247          5.7         2.9          4.2         1.3 versicolor
#> 248          6.2         2.9          4.3         1.3 versicolor
#> 249          5.1         2.5          3.0         1.1 versicolor
#> 250          5.7         2.8          4.1         1.3 versicolor
#> 251          6.3         3.3          6.0         2.5  virginica
#> 252          5.8         2.7          5.1         1.9  virginica
#> 253          7.1         3.0          5.9         2.1  virginica
#> 254          6.3         2.9          5.6         1.8  virginica
#> 255          6.5         3.0          5.8         2.2  virginica
#> 256          7.6         3.0          6.6         2.1  virginica
#> 257          4.9         2.5          4.5         1.7  virginica
#> 258          7.3         2.9          6.3         1.8  virginica
#> 259          6.7         2.5          5.8         1.8  virginica
#> 260          7.2         3.6          6.1         2.5  virginica
#> 261          6.5         3.2          5.1         2.0  virginica
#> 262          6.4         2.7          5.3         1.9  virginica
#> 263          6.8         3.0          5.5         2.1  virginica
#> 264          5.7         2.5          5.0         2.0  virginica
#> 265          5.8         2.8          5.1         2.4  virginica
#> 266          6.4         3.2          5.3         2.3  virginica
#> 267          6.5         3.0          5.5         1.8  virginica
#> 268          7.7         3.8          6.7         2.2  virginica
#> 269          7.7         2.6          6.9         2.3  virginica
#> 270          6.0         2.2          5.0         1.5  virginica
#> 271          6.9         3.2          5.7         2.3  virginica
#> 272          5.6         2.8          4.9         2.0  virginica
#> 273          7.7         2.8          6.7         2.0  virginica
#> 274          6.3         2.7          4.9         1.8  virginica
#> 275          6.7         3.3          5.7         2.1  virginica
#> 276          7.2         3.2          6.0         1.8  virginica
#> 277          6.2         2.8          4.8         1.8  virginica
#> 278          6.1         3.0          4.9         1.8  virginica
#> 279          6.4         2.8          5.6         2.1  virginica
#> 280          7.2         3.0          5.8         1.6  virginica
#> 281          7.4         2.8          6.1         1.9  virginica
#> 282          7.9         3.8          6.4         2.0  virginica
#> 283          6.4         2.8          5.6         2.2  virginica
#> 284          6.3         2.8          5.1         1.5  virginica
#> 285          6.1         2.6          5.6         1.4  virginica
#> 286          7.7         3.0          6.1         2.3  virginica
#> 287          6.3         3.4          5.6         2.4  virginica
#> 288          6.4         3.1          5.5         1.8  virginica
#> 289          6.0         3.0          4.8         1.8  virginica
#> 290          6.9         3.1          5.4         2.1  virginica
#> 291          6.7         3.1          5.6         2.4  virginica
#> 292          6.9         3.1          5.1         2.3  virginica
#> 293          5.8         2.7          5.1         1.9  virginica
#> 294          6.8         3.2          5.9         2.3  virginica
#> 295          6.7         3.3          5.7         2.5  virginica
#> 296          6.7         3.0          5.2         2.3  virginica
#> 297          6.3         2.5          5.0         1.9  virginica
#> 298          6.5         3.0          5.2         2.0  virginica
#> 299          6.2         3.4          5.4         2.3  virginica
#> 300          5.9         3.0          5.1         1.8  virginica